+++ src/client/pages/SignIn.tsx
@@ -13,7 +13,7 @@ import { gridItemSignInAndRegistration } from '@/client/styles/Grid';
 import { from, textSans, border, space } from '@guardian/source-foundations';
 import { Divider } from '@guardian/source-react-components-development-kitchen';
 import { CaptchaErrors, SignInErrors } from '@/shared/model/Errors';
-import { EmailInput } from '@/client/components/EmailInput';
+import { EmailInput } from '../components/EmailInput.importable';
 import { buildUrlWithQueryParams } from '@/shared/lib/routeUtils';
 import { QueryParams } from '@/shared/model/QueryParams';
 import { DetailedRecaptchaError } from '@/client/components/DetailedRecaptchaError';
@@ -24,6 +24,7 @@ import locations from '@/shared/lib/locations';
 import { RefTrackingFormFields } from '@/client/components/RefTrackingFormFields';
 import { trackFormFocusBlur, trackFormSubmit } from '@/client/lib/ophan';
 import { logger } from '@/client/lib/clientSideLogger';
+import { Islet } from '@/client/components/Islet';
 
 export type SignInProps = {
   returnUrl?: string;
@@ -221,7 +222,9 @@ export const SignIn = ({
           <RecaptchaElement id="signin-recaptcha" />
           <CsrfFormField />
           <RefTrackingFormFields />
-          <EmailInput defaultValue={email} />
+          <Islet type="component" deferUntil="idle">
+            <EmailInput defaultValue={email} />
+          </Islet>
           <div css={passwordInput}>
             <PasswordInput label="Password" autoComplete="current-password" />
           </div>
